loopback: a local client/server for testing...
This doesn't do anything yet. I'm playing around with fanotify which is
a successor (sort of) to inotify. IT DOES NOT WORK YET. I'm getting a
resource busy error and do not know why.

Right now I have it set to monitor the actual import device, but the
best would be to have it monitor the entire netchar directory.