netchar/loopback: first go at communication
this isn't quite working and will probably cause the need to restart
your computer if you access /dev/netchar/imp*. something needs to  be
explicitly set, I think in the module.

it's also pretty messy (understandably) and needs some cleanup.
furthermore, it's pretty inefficient and that needs to be looked at.