netchar/loopback: check in open/release, some read
It's time to just check in some work. Seems like open/close are working
ok. Read is working in part, but need to test more. Write should be easy
once read is done.

Write now we're just polling through reads/writes to a char device that
servers to pass a struct back and forth. Shared memory would be a lot
more efficient, but it lacks synchronization. Netlink sockets would
probably be the best solution.